+++ local.go
@@ -1,4 +1,3 @@
-// Package k6build defines a service for building k8 binaries
 package k6build
 
 import (
@@ -12,64 +11,6 @@ import (
 	"github.com/grafana/k6catalog"
 	"github.com/grafana/k6foundry"
 )
-
-const (
-	k6Dep = "k6"
-)
-
-// Dependency contains the properties of a k6 dependency.
-type Dependency struct {
-	// Name is the name of the dependency.
-	Name string `json:"name,omitempty"`
-	// Constraints contains the version constraints of the dependency.
-	Constraints string `json:"constraints,omitempty"`
-}
-
-// Module defines an artifact dependency
-type Module struct {
-	Path    string `json:"path,omitempty"`
-	Version string `json:"vesion,omitempty"`
-}
-
-// Artifact defines a binary that can be downloaded
-// TODO: add metadata (e.g. list of dependencies, checksum, date compiled)
-type Artifact struct {
-	ID string `json:"id,omitempty"`
-	// URL to fetch the artifact's binary
-	URL string `json:"url,omitempty"`
-	// list of dependencies
-	Dependencies map[string]string `json:"dependencies,omitempty"`
-	// platform
-	Platform string `json:"platform,omitempty"`
-	// binary checksum (sha256)
-	Checksum string `json:"checksum,omitempty"`
-}
-
-// BuildService defines the interface of a build service
-type BuildService interface {
-	// Build returns a k6 Artifact given its dependencies and version constrain
-	Build(ctx context.Context, platform string, k6Constrains string, deps []Dependency) (Artifact, error)
-}
-
-// implements the BuildService interface
-type buildsrv struct {
-	catalog k6catalog.Catalog
-	builder k6foundry.Builder
-	cache   Cache
-}
-
-// NewBuildService creates a build service
-func NewBuildService(
-	catalog k6catalog.Catalog,
-	builder k6foundry.Builder,
-	cache Cache,
-) BuildService {
-	return &buildsrv{
-		catalog: catalog,
-		builder: builder,
-		cache:   cache,
-	}
-}
 
 // LocalBuildServiceConfig defines the configuration for a Local build service
 type LocalBuildServiceConfig struct {
