+++ local_test.go
@@ -3,96 +3,15 @@ package k6build
 import (
 	"context"
 	"errors"
-	"fmt"
-	"net/http/httptest"
 	"testing"
 
 	"github.com/grafana/k6catalog"
-	"github.com/grafana/k6foundry"
-	"github.com/grafana/k6foundry/pkg/testutils/goproxy"
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
 )
 
 func dependencyComp(a, b Module) bool { return a.Path < b.Path }
-
-func setupBuildService(cacheDir string) (BuildService, error) {
-	modules := []struct {
-		path    string
-		version string
-		source  string
-	}{
-		{
-			path:    "go.k6.io/k6",
-			version: "v0.1.0",
-			source:  "testdata/deps/k6",
-		},
-		{
-			path:    "go.k6.io/k6",
-			version: "v0.2.0",
-			source:  "testdata/deps/k6",
-		},
-		{
-			path:    "go.k6.io/k6ext",
-			version: "v0.1.0",
-			source:  "testdata/deps/k6ext",
-		},
-		{
-			path:    "go.k6.io/k6ext",
-			version: "v0.2.0",
-			source:  "testdata/deps/k6ext",
-		},
-		{
-			path:    "go.k6.io/k6ext2",
-			version: "v0.1.0",
-			source:  "testdata/deps/k6ext2",
-		},
-	}
-
-	// creates a goproxy that serves the given modules
-	proxy := goproxy.NewGoProxy()
-	for _, m := range modules {
-		err := proxy.AddModVersion(m.path, m.version, m.source)
-		if err != nil {
-			return nil, fmt.Errorf("setup %w", err)
-		}
-	}
-
-	goproxySrv := httptest.NewServer(proxy)
-
-	opts := k6foundry.NativeBuilderOpts{
-		GoOpts: k6foundry.GoOpts{
-			CopyGoEnv: true,
-			Env: map[string]string{
-				"GOPROXY":   goproxySrv.URL,
-				"GONOPROXY": "none",
-				"GOPRIVATE": "go.k6.io",
-				"GONOSUMDB": "go.k6.io",
-			},
-			TmpCache: true,
-		},
-	}
-
-	builder, err := k6foundry.NewNativeBuilder(context.Background(), opts)
-	if err != nil {
-		return nil, fmt.Errorf("setting up test builder %w", err)
-	}
-
-	catalog, err := k6catalog.NewCatalogFromJSON("testdata/catalog.json")
-	if err != nil {
-		return nil, fmt.Errorf("setting up test builder %w", err)
-	}
-
-	cache, err := NewFileCache(cacheDir)
-	if err != nil {
-		return nil, fmt.Errorf("creating temp cache %w", err)
-	}
-
-	buildsrv := NewBuildService(catalog, builder, cache)
-
-	return buildsrv, nil
-}
 
 func TestDependencyResolution(t *testing.T) {
 	t.Parallel()
@@ -153,7 +72,12 @@ func TestDependencyResolution(t *testing.T) {
 		t.Run(tc.title, func(t *testing.T) {
 			t.Parallel()
 
-			buildsrv, err := setupBuildService(t.TempDir())
+			buildsrv, err := SetupTestLocalBuildService(
+				LocalBuildServiceConfig{
+					CacheDir: t.TempDir(),
+					Catalog:  "testdata/catalog.json",
+				},
+			)
 			if err != nil {
 				t.Fatalf("test setup %v", err)
 			}
@@ -184,8 +108,12 @@ func TestDependencyResolution(t *testing.T) {
 
 func TestIdempotentBuild(t *testing.T) {
 	t.Parallel()
-
-	buildsrv, err := setupBuildService(t.TempDir())
+	buildsrv, err := SetupTestLocalBuildService(
+		LocalBuildServiceConfig{
+			CacheDir: t.TempDir(),
+			Catalog:  "testdata/catalog.json",
+		},
+	)
 	if err != nil {
 		t.Fatalf("test setup %v", err)
 	}
